@@ -8,9 +8,19 @@ namespace stellar_dotnet_sdk_test.Generators
     {
         public static Arbitrary<string> Generate()
         {
-            var regexGenerator = new Xeger("([a-zA-Z0-9]){5,12}");
+            var codeLength = Gen.Choose(5, 12);
+            var lowerCaseLetter = Gen.Choose('a', 'z');
+            var upperCaseLetter = Gen.Choose('A', 'Z');
+            var number = Gen.Choose('0', '9');
+            var validCharacter =
+                Gen.OneOf(lowerCaseLetter, upperCaseLetter, number)
+                    .Select(character => (char)character);
 
-            return Gen.Sized(assetCodeLength => Gen.Elements(Enumerable.Range(1, assetCodeLength).Select(_ => regexGenerator.Generate()))).ToArbitrary();
+            return
+                codeLength
+                    .SelectMany(length => Gen.ListOf(length, validCharacter))
+                    .Select(characters => new string(characters.ToArray()))
+                    .ToArbitrary();
         }
     }
 }
