@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Collections.Generic;
 using System.IO;
 using FluentAssertions;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
@@ -79,7 +80,7 @@ namespace stellar_dotnet_sdk_test.responses
             Assert.AreEqual(account.Flags.AuthImmutable, true);
             Assert.AreEqual(account.Flags.AuthClawback, true);
 
-            account.Balances
+            ((IEnumerable<Balance>)account.Balances)
                 .Should().HaveCount(3);
 
             Assert.AreEqual(account.Balances[0].AssetType, "credit_alphanum4");
@@ -89,6 +90,10 @@ namespace stellar_dotnet_sdk_test.responses
             Assert.IsInstanceOfType(asset, typeof(AssetTypeCreditAlphaNum));
             Assert.AreEqual(asset.Code, "ABC");
             Assert.AreEqual(asset.Issuer, "GCRA6COW27CY5MTKIA7POQ2326C5ABYCXODBN4TFF5VL4FMBRHOT3YHU");
+
+            ((IEnumerable<Balance>)account.Balances)
+                .Should().HaveCount(3);
+
             Assert.AreEqual(account.Balances[0].BalanceString, "1001.0000000");
             Assert.AreEqual(account.Balances[0].Limit, "12000.4775807");
             Assert.AreEqual(account.Balances[0].BuyingLiabilities, "100.1234567");
@@ -105,7 +110,7 @@ namespace stellar_dotnet_sdk_test.responses
             Assert.AreEqual(account.Balances[1].Limit, null);
 
             // liquidity pool balance
-            account.Balances[2].Asset
+            ((object)account.Balances[2].Asset)
                 .Should().BeNull();
 
             account.Balances[2].AssetType
