@@ -185,7 +185,7 @@ namespace stellar_dotnet_sdk
                     responseString = await response.Content.ReadAsStringAsync();
                 }
 
-                throw new ConnectionErrorException($"Status code ({response.StatusCode}) is not success.{ (!string.IsNullOrEmpty(responseString) ? " Content: " + responseString : "") }");
+                throw new ConnectionErrorException($"Status code ({response.StatusCode}) is not success.{(!string.IsNullOrEmpty(responseString) ? " Content: " + responseString : "")}");
             }
 
             if (response.Content != null)
