@@ -38,6 +38,6 @@ namespace stellar_dotnet_sdk.responses.effects
         [JsonProperty(PropertyName = "asset_issuer")]
         public string AssetIssuer { get; private set; }
 
-        public Asset Asset => Asset.CreateNonNativeAsset(AssetType, AssetIssuer, AssetCode);
+        public Asset Asset => Asset.Create(AssetType, AssetCode, AssetIssuer);
     }
 }
